Own the ghost entry handlers in useBlockProps
The context carries the ghost block; useBlockProps derives the entry
handlers where they are applied. Also restore the guard that keeps the
classic appender from rendering next to a ghost.
